Split Devin provider tests by responsibility
